hw/display/apple-gfx-ml: consume cursor sentinel with signed coordinates
anomaly_id: VIS102_CURSOR_SENTINEL_SIGNEDNESS_VISIBLE_BACKEND

root cause: apple-gfx-ml split packed cursor position into unsigned halves,

so sentinel 0xffffffff became 65535,65535 and visible backends treated hidden

cursor doorbells as real pointer warps.

fix: reinterpret packed cursor halves as signed 16-bit display coordinates

before dpy_mouse_set() and store wrapper cursor state in signed fields.

evidence: apple-gfx.m consumes PGDisplay cursor coordinates; gtk and sdl

visible backends turn dpy_mouse_set() into real pointer warps; A/B/A visible

guest runs reproduce corruption, remove it under this change, and restore it

after exact revert.

+++ hw/display/apple-gfx-ml.c
@@ -863,8 +863,8 @@ static void apple_gfx_ml_update_cursor(AppleGfxMLState *s, uint32_t display_id)
 
     if (s->qmu_dev &&
         qmu_get_display_cursor_position(s->qmu_dev, display_id, &packed) == QMU_OK) {
-        s->cursor_x = packed & 0xffffu;
-        s->cursor_y = (packed >> 16) & 0xffffu;
+        s->cursor_x = (int16_t)(packed & 0xffffu);
+        s->cursor_y = (int16_t)((packed >> 16) & 0xffffu);
     }
     dpy_mouse_set(s->con, s->cursor_x, s->cursor_y, s->cursor_show);
 }
